+++ www/go/modules/community/tasks/model/Category.php
@@ -9,6 +9,7 @@ namespace go\modules\community\tasks\model;
 use go\core\db\Criteria;
 use go\core\jmap\Entity;
 use go\core\model\Acl;
+use go\core\model\Module;
 use go\core\orm\Filters;
 use go\core\orm\Mapping;
 
@@ -37,7 +38,12 @@ class Category extends Entity {
 
 	public function getPermissionLevel(): int
 	{
-		return ($this->ownerId === go()->getUserId() || \go\core\model\Module::findByName('community', 'tasks')->hasPermissionLevel(50)) ?  Acl::LEVEL_MANAGE : 0; // validate will make sure global categories aren't changed if no permission
+		return (
+			$this->ownerId === go()->getUserId()
+			||
+			Module::findByName('community', 'tasks')
+				->getUserRights()
+				->mayChangeCategories ? Acl::LEVEL_MANAGE : 0);
 	}
 
 	public function internalValidate()
@@ -45,7 +51,7 @@ class Category extends Entity {
 		if($this->isNew() && !$this->isModified('ownerId')) {
 			$this->ownerId = go()->getUserId();
 		}
-		if ($this->ownerId !== go()->getUserId() && !\go\core\model\Module::findByName('community', 'tasks')->hasPermissionLevel(50))
+		if ($this->ownerId !== go()->getUserId() && !Module::findByName('community', 'tasks')->hasPermissionLevel(50))
 			$this->setValidationError('ownerId', go()->t("You need manage permission to create global categories"));
 		return parent::internalValidate();
 	}